littlelight

+++ weenix/kernel/api/access.c
@@ -158,16 +158,16 @@ int range_perm(struct proc *p, const void *avaddr, size_t len, int perm)
 {
         /*NOT_YET_IMPLEMENTED("VM: range_perm");*/
 	vmarea_t *myVmarea;
-	size_t tmp_len=0;
+	uint32_t tmp_len=0;
         list_iterate_begin(&p->p_vmmap->vmm_list,myVmarea,vmarea_t,vma_plink){
-                if(PN_TO_ADDR(myVmarea->vma_start)<= (avaddr+tmp_len) && PN_TO_ADDR(myVmarea->vma_end)>= (avaddr+tmp_len)){
+                if((uint32_t)PN_TO_ADDR(myVmarea->vma_start)<= ((uint32_t)avaddr+tmp_len) && (uint32_t)PN_TO_ADDR(myVmarea->vma_end)>= ((uint32_t)avaddr+tmp_len)){
                         if((myVmarea->vma_prot&perm)!=perm)
                                 return 0;
                         
-			if(PN_TO_ADDR(myVmarea->vma_end)-(avaddr+tmp_len)+1>=len)
+			if((uint32_t)PN_TO_ADDR(myVmarea->vma_end)-((uint32_t)avaddr+tmp_len)+1>=len)
 				return 1;
 			else{
-				tmp_len = PN_TO_ADDR(myVmarea->vma_end)+1-(avaddr+tmp_len)+1;
+				tmp_len = (uint32_t)PN_TO_ADDR(myVmarea->vma_end)+1-((uint32_t)avaddr+tmp_len)+1;
 				len = len - tmp_len;
 			}	
                 }
